requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html

# --- FastAPI stack ---
fastapi==0.110.0
uvicorn[standard]==0.27.1
jinja2==3.1.3
python-multipart==0.0.9

# --- PostgreSQL & ORM ---
psycopg2-binary==2.9.9

# --- HTTP & API ---
requests==2.31.0
sickle==0.7.0

# --- Asynchrone & tâches ---
celery[redis,timezone]==5.3.6

# --- NLP & ML ---
transformers==4.39.1
torch==2.2.1+cpu
numpy==1.26.4

# --- PDF parsing ---
PyMuPDF==1.23.9

# --- XML parsing (GROBID) ---
lxml==4.9.3

feedparser>=6.0.0,<7.0.0
starlette>=0.36.3,<0.37.0
pydantic>=2.11.2,<3.0.0

httpx>=0.24.1,<1.0.0
beautifulsoup4>=4.11.1

huggingface_hub[hf_xet]==0.30.2